feat : "OrderViewController" 이름변경 및 Alert 로직 구현
